28/04/2018 Migrating database on MySql completed

+++ src/java/com/raffaelemancino/ricettario/core/view/IngredientiPerRicetta.java
@@ -23,7 +23,7 @@ package com.raffaelemancino.ricettario.core.view;
  */
 public class IngredientiPerRicetta
 {
-    public Integer qt;
+    public Float qt;
     public String unit;
     public String name;
 }
